requirements.txt
asgiref==3.11.1
dataclasses==0.6
Django==6.0.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djongo==1.3.6
dnspython==2.8.0
legacy-cgi==2.6.4
PyJWT==2.11.0
pymongo==3.13.0
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
